+++ homeassistant/util/yaml/constructor.py
@@ -1,0 +1,1 @@
+"""Custom yaml constructor."""
